cleanup

+++ release.sh
@@ -1,20 +1,48 @@
 #!/bin/bash -e
 
+
 # determine full version
-BASE_VERSION="$(cat 'version' | xargs).${BUILD_NUMBER:-local}"
-GIT_SHA=$(git rev-parse --short HEAD)
-DIRTY=$([[ -z $(git status -s) ]] || echo '-dirty')
-VERSION=${BASE_VERSION}-${GIT_SHA}${DIRTY}
+VER_LONG=$(git describe --tags --long | cut -c2-)
+echo "VER_LONG: ${VER_LONG}"
+# note that the first character "v" had to be stripped, debian requires version numbers to start with digit
+# (git convention is to use a "v" as prefix in the version number)
+# example (without "v"): 2.1.1-2-g146e8fc
+
+
+VER_SHORT=$(echo ${VER_LONG} | cut -d '-' -f 1)
+echo "VER_SHORT: ${VER_SHORT}"
+# example: 2.1.1
+
+REL_COMMIT_COUNT=$(echo ${VER_LONG} | cut -d '-' -f 2)
+echo "REL_COMMIT_COUNT: ${REL_COMMIT_COUNT}"
+# if this is not 0, do not do a release
+# example: 2
+
+
+if [[ ( "${REL_COMMIT_COUNT}_" != "0_" ) && "$1_" != "--force_" ]] ; then
+    echo ""
+    echo "Error:"
+    echo "  The current git commit has not been tagged. Please create a new tag first to ensure a proper unique version number."
+    echo "  Use --force to ignore error (for debugging only)"
+    echo ""
+    exit 1
+fi
+
+
+
+VER_HASH=$(echo ${VER_LONG} | cut -d '-' -f 1)
+echo "VER_SHASH: ${VER_HASH}"
+
 
 # Build the registration debian package
 BASEDIR=/tmp/reg
-NAME=sagebk-registration
+NAME=waggle-registration
 ARCH=all
 
 mkdir -p ${BASEDIR}/DEBIAN
 cat > ${BASEDIR}/DEBIAN/control <<EOL
 Package: ${NAME}
-Version: ${VERSION}
+Version: ${VER_LONG}
 Maintainer: sagecontinuum.org
 Description: Register with Beehive server
 Architecture: ${ARCH}
@@ -26,32 +54,8 @@ cp -p deb/reg/prerm ${BASEDIR}/DEBIAN/
 
 mkdir -p ${BASEDIR}/etc/systemd/system
 mkdir -p ${BASEDIR}/usr/bin
-cp -p ROOTFS/etc/systemd/system/sagebk-registration.service ${BASEDIR}/etc/systemd/system/
-cp -p ROOTFS/usr/bin/sagebk-registration ${BASEDIR}/usr/bin
+cp -p waggle-registration.service ${BASEDIR}/etc/systemd/system/
+cp -p waggle-registration.py ${BASEDIR}/usr/bin
 
-dpkg-deb --root-owner-group --build ${BASEDIR} "${NAME}_${VERSION}_${ARCH}.deb"
-
-# Build the reverse tunnel debian package
-BASEDIR=/tmp/reverse
-NAME=sagebk-reverse-tunnel
-ARCH=all
-
-mkdir -p ${BASEDIR}/DEBIAN
-cat > ${BASEDIR}/DEBIAN/control <<EOL
-Package: ${NAME}
-Version: ${VERSION}
-Maintainer: sagecontinuum.org
-Description: Establish reverse SSH tunnel to Beehive
-Architecture: ${ARCH}
-Priority: optional
-EOL
-
-cp -p deb/reverse/postinst ${BASEDIR}/DEBIAN/
-cp -p deb/reverse/prerm ${BASEDIR}/DEBIAN/
-
-mkdir -p ${BASEDIR}/etc/systemd/system
-mkdir -p ${BASEDIR}/usr/bin
-cp -p ROOTFS/etc/systemd/system/sagebk-reverse-tunnel.service ${BASEDIR}/etc/systemd/system/
-cp -p ROOTFS/usr/bin/sagebk-reverse-tunnel ${BASEDIR}/usr/bin
-
-dpkg-deb --root-owner-group --build ${BASEDIR} "${NAME}_${VERSION}_${ARCH}.deb"
+set -x
+dpkg-deb --root-owner-group --build ${BASEDIR} "${NAME}_${VER_SHORT}_${ARCH}.deb"
